// generated/src/app/api-models/reason-code-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ReasonCodeMaster {
  reasonCode: string;
  reasonCodeType: string;
  description: string;
  autoAuditLocCod: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  uncleanFlag: string;
  userDefined1: string;
  userDefined2: string;
  userDate1: Date;
  userDate2: Date;
  description2: string;
  calcDescription: string;
  eobCalcOption: string;
  descInd: string;
  rncCode: string;
}